requirements.txt
setuptools>=57.0.0
numpy>=1.21.1
scipy>=1.7.1
spatialmath-python>=0.10